// src/dsep/dsep-modules.enum.ts
export enum DSEPModulesEnum {
    SCHOLARSHIPS_GRANTS = "dsep:scholarships",
    COURSES_TRAINING = "dsep:courses",
    JOBS_INTERNSHIPS = "dsep:jobs",
    TUTORING_MENTORSHIPS = "dsep:mentoring",
    STUDENT_CONNECT = "dsep:student-connect",
    ONEST_WORK_OPPORTUNITIES = "onest:work-opportunities",
    ONEST_FINANCIAL_SUPPORT = "onest:financial-support",
    ONEST_LEARNING_EXPERIENCES = "onest:learning-experiences",
    ONEST_EXPERT_CONNECT = "onest:expert-connect",
}